csapp.c

+++ run_source/my_web_server/version4/cd_io_wrap.h
@@ -6,7 +6,7 @@ ssize_t cd_readn(int fd, void *vptr, size_t n);
 ssize_t cd_writen(int fd, const void *vptr, size_t n);
 ssize_t cd_my_read(int fd, char *ptr);
 ssize_t cd_read_line(int fd, void *vptr, size_t maxlen);
-int cd_get_line(int sock, char *buf, int size);
+ssize_t cd_get_line(int sock, char *buf, int size);
 int cd_close(int fd);
 int cd_port_reuse(int fd);
 #endif
